+++ src/components/navbar/ConnectedWallets.tsx
@@ -51,9 +51,12 @@ const ConnectedWallets = () => {
     <>
       <Opacity
         level="medium"
-        className="ml-auto flex min-h-[32px] min-w-[32px] cursor-pointer items-center justify-center gap-1.5 rounded-[24px] p-2 transition-all duration-300 hover:bg-opacity-80 sm:min-h-[40px] sm:min-w-[40px] sm:gap-2 sm:p-3"
+        className="relative z-0 ml-auto flex min-h-[32px] min-w-[32px] cursor-pointer items-center justify-center gap-1.5 overflow-hidden rounded-[24px] p-2 transition-all duration-300 hover:bg-opacity-80 sm:min-h-[40px] sm:min-w-[40px] sm:gap-2 sm:p-3"
         onClick={handleAddressClick}
       >
+        {!!pendingOrdersCount && (
+          <div className="navbar-shine absolute left-0 top-0 -z-10 h-full w-full translate-x-[-25%] scale-150" />
+        )}
         <WalletIcon className="h-4 w-4 sm:h-5 sm:w-5" />
         {address && (
           <img
@@ -84,8 +87,8 @@ const ConnectedWallets = () => {
           />
         )}
         {pendingOrdersCount ? (
-          <div className="flex h-5 w-5 items-center justify-center rounded-full border border-rose p-2">
-            <Typography size="h5" weight="bold" className="h-4 !text-rose">
+          <div className="flex h-5 w-5 items-center justify-center rounded-full bg-rose p-2">
+            <Typography size="h5" weight="bold" className="h-4 !text-white">
               {pendingOrdersCount}
             </Typography>
           </div>